HW3

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcMealRepositoryImpl.java
@@ -56,26 +56,31 @@ public class JdbcMealRepositoryImpl implements MealRepository {
             Number newKey = insertMeal.executeAndReturnKey(map);
             meal.setId(newKey.intValue());
         } else {
-            namedParameterJdbcTemplate.update(
-                    "UPDATE meals SET user_id=:user_id, datetime=:dateTime, description=:description, calories=:calories WHERE id=:id", map);
+            if (namedParameterJdbcTemplate.update("" +
+                    "UPDATE meals " +
+                    "   SET description=:description, calories=:calories, datetime=:datetime " +
+                    " WHERE id=:id AND user_id=:user_id"
+                    , map) == 0) {
+                return null;
+            }
         }
         return meal;
     }
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update("DELETE FROM meals WHERE id=?", id) != 0;
+        return jdbcTemplate.update("DELETE FROM meals WHERE id=? AND user_id=?", id, userId) != 0;
     }
 
     @Override
     public Meal get(int id, int userId) {
-        List<Meal> meals = jdbcTemplate.query("SELECT * FROM meals WHERE id=?", ROW_MAPPER, id);
+        List<Meal> meals = jdbcTemplate.query("SELECT * FROM meals WHERE id=? AND user_id=?", ROW_MAPPER, id, userId);
         return DataAccessUtils.singleResult(meals);
     }
 
     @Override
     public List<Meal> getAll(int userId) {
-        return jdbcTemplate.query("SELECT * FROM meals ORDER BY datetime", ROW_MAPPER);
+        return jdbcTemplate.query("SELECT * FROM meals WHERE user_id ORDER BY datetime DESC", ROW_MAPPER, userId);
     }
 
     @Override
@@ -83,8 +88,8 @@ public class JdbcMealRepositoryImpl implements MealRepository {
         Objects.requireNonNull(startDate);
         Objects.requireNonNull(endDate);
 
-        return jdbcTemplate.query("SELECT * FROM meals WHERE user_id=? ORDER BY datetime", ROW_MAPPER, userId).stream()
-                .filter(p -> DateTimeUtil.isBetween(p.getDateTime(), startDate, endDate))
-                .collect(Collectors.toList());
+        return jdbcTemplate.query(
+                "SELECT * FROM meals WHERE user_id=?  AND date_time BETWEEN  ? AND ? ORDER BY date_time DESC",
+                ROW_MAPPER, userId, startDate, endDate);
     }
 }
